+++ src/certsf/methods/__init__.py
@@ -1,5 +1,6 @@
 """Custom certified methods."""
 
 from .stirling import stirling_loggamma, stirling_loggamma_shifted
+from .loggamma_auto import certified_auto_loggamma
 
-__all__ = ["stirling_loggamma", "stirling_loggamma_shifted"]
+__all__ = ["certified_auto_loggamma", "stirling_loggamma", "stirling_loggamma_shifted"]